Fix compilation of neighborhood class
The previous commit 7f29ca2 built
under GCC but broke compiling
under Visual Studio. GCC disallowed
the static constexpr variable in a
function, but MSVC was more strict
about the order in which templates
were defined.

The neighborhood classes are not
split into a lower level (MoveData and
NeighborhoodData) and an upper
level (Move and Neighborhood)
establishing an ordering of template
definitions that satisfies MSVC.

Also updated the python scripts due
to problems checking paths with
backslashes.